+++ packages/components/src/dropdown-menu-v2-ariakit/index.tsx
@@ -26,6 +26,11 @@ import { SVG, Circle } from '@wordpress/primitives';
 import { useContextSystem, contextConnect } from '../context';
 import type { WordPressComponentProps } from '../context';
 import Icon from '../icon';
+import { useSlot } from '../slot-fill';
+import {
+	SLOT_NAME as POPOVER_DEFAULT_SLOT_NAME,
+	slotNameContext,
+} from '../popover';
 import type {
 	DropdownMenuContext as DropdownMenuContextType,
 	DropdownMenuProps,
@@ -182,6 +187,9 @@ const UnconnectedDropdownMenu = (
 		shift,
 		modal = true,
 
+		// Other props
+		slotName: slotNameProp = POPOVER_DEFAULT_SLOT_NAME,
+
 		// From internal components context
 		variant,
 
@@ -270,6 +278,11 @@ const UnconnectedDropdownMenu = (
 		[ computedDirection ]
 	);
 
+	// Render the portal in the default slot used by the legacy Popover component.
+	const slotName = useContext( slotNameContext ) || slotNameProp;
+	const slot = useSlot( slotName );
+	const portalContainer = slot.ref?.current;
+
 	return (
 		<>
 			{ /* Menu trigger */ }
@@ -305,6 +318,7 @@ const UnconnectedDropdownMenu = (
 				wrapperProps={ wrapperProps }
 				hideOnEscape={ hideOnEscape }
 				unmountOnHide
+				portalElement={ portalContainer }
 			>
 				<DropdownMenuContext.Provider value={ contextValue }>
 					{ children }